Clever method with pointers.
Before the app kept crashing either because the flag Qt::DElete_ON_Close
was activated and the window would delete itself when closing, but cause
problems on pointer reassignement.

Solved by setting pointers by default to 0.
Then when MotionWind is activated, it checks to see if it's a null
pointer or not and acts accordingly.

To limit the amount memory,  a delete method is called on TimeLapse
window when MotionWind is initiated, and vice versa.

Why is this effective? If you choose a window, then go back from it,
then go back into it agin -- the pointer is not lost and it calls the
same object. However if you choose to try a different window, it deletes
the old.